+++ source/core/src/main/com/csse3200/game/areas/TestGameArea.java
@@ -1,5 +1,6 @@
 package com.csse3200.game.areas;
 
+import com.csse3200.game.areas.levels.LevelFactory;
 import com.csse3200.game.entities.Entity;
 
 /**
