+++ source/core/src/main/com/csse3200/game/areas/levels/Level.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.areas;
+package com.csse3200.game.areas.levels;
 
 import com.csse3200.game.entities.Room;
 import org.slf4j.Logger;
