+++ source/core/src/main/com/csse3200/game/areas/levels/LevelMap.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.areas;
+package com.csse3200.game.areas.levels;
 
 import com.csse3200.game.areas.Generation.MapGenerator;
 import com.csse3200.game.files.FileLoader;
